Simplify legal entity overview
