+++ apps/web/src/components/dashboard-kpis.tsx
@@ -3,13 +3,7 @@ import type { ReactNode } from "react"
 import { Box, Card, Divider, Group, Stack, Table, Text } from "@mantine/core"
 import { useQuery } from "@tanstack/react-query"
 
-import {
-  getHairAssignedStatsForDate,
-  getHairAssignedThroughSaleStatsForDate,
-  getTransactionStatsForDate,
-  type HairMonthlyBreakdown,
-  type TransactionMonthlyByCurrency,
-} from "@/functions/dashboard"
+import { getTransactionStatsForDate, type TransactionMonthlyByCurrency } from "@/functions/dashboard"
 import { type Currency, formatMinor } from "@/lib/currency"
 import { dashboardKeys } from "@/lib/query-keys"
 
@@ -20,21 +14,10 @@ export function DashboardKpis({ year, legalEntityId }: { year: number; legalEnti
     queryKey: dashboardKeys.transactionStats(year, legalEntityId),
     queryFn: () => getTransactionStatsForDate({ data: { year, legalEntityId: legalEntityId || undefined } }),
   })
-  const hairAptQuery = useQuery({
-    queryKey: [...dashboardKeys.hairAssignedStats(year), legalEntityId],
-    queryFn: () => getHairAssignedStatsForDate({ data: { year, legalEntityId: legalEntityId || undefined } }),
-  })
-  const hairSaleQuery = useQuery({
-    queryKey: [...dashboardKeys.hairSaleStats(year), legalEntityId],
-    queryFn: () =>
-      getHairAssignedThroughSaleStatsForDate({ data: { year, legalEntityId: legalEntityId || undefined } }),
-  })
 
   return (
     <Stack gap="md">
       <TransactionsCard data={txQuery.data ?? null} />
-      <HairCard title="Hair assigned during appointments" data={hairAptQuery.data ?? null} />
-      <HairCard title="Hair assigned during sale" data={hairSaleQuery.data ?? null} />
     </Stack>
   )
 }
@@ -97,70 +80,6 @@ function TransactionsCard({ data }: { data: TransactionMonthlyByCurrency[] | nul
   )
 }
 
-function HairCard({ title, data }: { title: string; data: HairMonthlyBreakdown | null }) {
-  return (
-    <KpiPanel title={title}>
-      {data === null ? (
-        <EmptyPanelText>Loading...</EmptyPanelText>
-      ) : (
-        <Box p="md">
-          <Table.ScrollContainer minWidth={680}>
-            <Table striped highlightOnHover>
-              <Table.Thead>
-                <Table.Tr>
-                  <Table.Th>Month</Table.Th>
-                  <Table.Th ta="right">Weight (g)</Table.Th>
-                  <Table.Th ta="right">Sold for</Table.Th>
-                  <Table.Th ta="right">Profit</Table.Th>
-                  <Table.Th ta="right">Price per gram</Table.Th>
-                </Table.Tr>
-              </Table.Thead>
-              <Table.Tbody>
-                {data.months.map((m) => (
-                  <Table.Tr key={m.month}>
-                    <Table.Td>{MONTH_NAMES[m.month - 1]}</Table.Td>
-                    <Table.Td ta="right" c={m.weight > 0 ? undefined : "dimmed"}>
-                      {m.weight}g
-                    </Table.Td>
-                    <Table.Td ta="right" c={m.soldFor > 0 ? undefined : "dimmed"}>
-                      {fmtCents(m.soldFor)}
-                    </Table.Td>
-                    <Table.Td ta="right" c={m.profit > 0 ? "teal" : m.profit < 0 ? "red" : "dimmed"}>
-                      {fmtCents(m.profit)}
-                    </Table.Td>
-                    <Table.Td ta="right" c={m.pricePerGram > 0 ? undefined : "dimmed"}>
-                      {fmtCents(m.pricePerGram)}
-                    </Table.Td>
-                  </Table.Tr>
-                ))}
-                <Table.Tr>
-                  <Table.Td fw={600}>Total</Table.Td>
-                  <Table.Td ta="right" fw={700}>
-                    {data.totals.weight}g
-                  </Table.Td>
-                  <Table.Td ta="right" fw={700}>
-                    {fmtCents(data.totals.soldFor)}
-                  </Table.Td>
-                  <Table.Td
-                    ta="right"
-                    fw={700}
-                    c={data.totals.profit > 0 ? "teal" : data.totals.profit < 0 ? "red" : undefined}
-                  >
-                    {fmtCents(data.totals.profit)}
-                  </Table.Td>
-                  <Table.Td ta="right" fw={700}>
-                    {fmtCents(data.totals.pricePerGramAvg)}
-                  </Table.Td>
-                </Table.Tr>
-              </Table.Tbody>
-            </Table>
-          </Table.ScrollContainer>
-        </Box>
-      )}
-    </KpiPanel>
-  )
-}
-
 function KpiPanel({ title, children }: { title: string; children: ReactNode }) {
   return (
     <Card padding={0}>
@@ -179,8 +98,4 @@ function EmptyPanelText({ children }: { children: ReactNode }) {
       {children}
     </Text>
   )
-}
-
-function fmtCents(cents: number) {
-  return `€${(cents / 100).toFixed(2)}`
 }
