+++ apps/web/src/functions/hair-orders.ts
@@ -1,4 +1,4 @@
-import { db, whereActiveLegalEntity } from "@prive-admin-tanstack/db"
+import { db } from "@prive-admin-tanstack/db"
 import { hairAssigned, hairOrder } from "@prive-admin-tanstack/db/schema/hair"
 import { createServerFn } from "@tanstack/react-start"
 import { eq } from "drizzle-orm"
@@ -9,11 +9,8 @@ import { requireAuthMiddleware } from "@/middleware/auth"
 
 export const getHairOrders = createServerFn({ method: "GET" })
   .middleware([requireAuthMiddleware])
-  .inputValidator(z.object({ legalEntityId: z.string().optional() }).optional())
-  .handler(async ({ data }) => {
-    const filter = data?.legalEntityId ? whereActiveLegalEntity(hairOrder.legalEntityId, data.legalEntityId) : undefined
+  .handler(async () => {
     return db.query.hairOrder.findMany({
-      where: filter,
       with: { createdBy: true, customer: true },
       orderBy: (hairOrder, { asc }) => [asc(hairOrder.uid)],
     })
@@ -49,7 +46,6 @@ export const createHairOrder = createServerFn({ method: "POST" })
         arrivedAt: data.arrivedAt ? String(data.arrivedAt) : null,
         status: data.status,
         customerId: data.customerId,
-        legalEntityId: data.legalEntityId,
         weightReceived: data.weightReceived,
         weightUsed: data.weightUsed,
         total: data.total,
@@ -72,7 +68,6 @@ export const updateHairOrder = createServerFn({ method: "POST" })
         weightReceived: data.weightReceived,
         weightUsed: data.weightUsed,
         total: data.total,
-        legalEntityId: data.legalEntityId,
       })
       .where(eq(hairOrder.id, data.id!))
       .returning()
