+++ apps/web/src/routes/_authenticated/hair-orders/$hairOrderId.tsx
@@ -8,7 +8,6 @@ import {
   Modal,
   NativeSelect,
   NumberInput,
-  Select,
   SimpleGrid,
   Skeleton,
   Stack,
@@ -31,8 +30,6 @@ import { HairAssignedTable, type HairAssignedRow } from "@/components/hair-assig
 import { PageHeader } from "@/components/page-header"
 import { Section } from "@/components/section"
 import { getHairOrder, recalculateHairOrderPrices, updateHairOrder } from "@/functions/hair-orders"
-import { listLegalEntities } from "@/functions/legal-entities"
-import { COUNTRY_FLAGS, type Country } from "@/lib/legal-entity"
 import { hairOrderKeys } from "@/lib/query-keys"
 
 export const Route = createFileRoute("/_authenticated/hair-orders/$hairOrderId")({
@@ -254,14 +251,11 @@ type EditHairOrderModalProps = {
     weightUsed: number
     total: number
     customerId: string
-    legalEntityId: string
   }
 }
 
 function EditHairOrderModal({ open, onOpenChange, hairOrder }: EditHairOrderModalProps) {
   const queryClient = useQueryClient()
-
-  const legalEntitiesQuery = useQuery({ queryKey: ["legal-entities"], queryFn: () => listLegalEntities() })
 
   const mutation = useMutation({
     mutationFn: (values: {
@@ -270,7 +264,6 @@ function EditHairOrderModal({ open, onOpenChange, hairOrder }: EditHairOrderModa
       status: "PENDING" | "COMPLETED"
       weightReceived: number
       total: number
-      legalEntityId: string
     }) =>
       updateHairOrder({
         data: {
@@ -282,7 +275,6 @@ function EditHairOrderModal({ open, onOpenChange, hairOrder }: EditHairOrderModa
           weightReceived: values.weightReceived,
           weightUsed: hairOrder.weightUsed,
           total: Math.round(values.total * 100),
-          legalEntityId: values.legalEntityId,
         },
       }),
     onSuccess: () => {
@@ -300,7 +292,6 @@ function EditHairOrderModal({ open, onOpenChange, hairOrder }: EditHairOrderModa
       status: (hairOrder.status === "COMPLETED" ? "COMPLETED" : "PENDING") as "PENDING" | "COMPLETED",
       weightReceived: hairOrder.weightReceived,
       total: hairOrder.total / 100,
-      legalEntityId: hairOrder.legalEntityId,
     },
   })
 
@@ -322,16 +313,6 @@ function EditHairOrderModal({ open, onOpenChange, hairOrder }: EditHairOrderModa
             <NumberInput label="Weight Received (g)" min={0} {...form.getInputProps("weightReceived")} />
             <NumberInput label="Total" min={0} decimalScale={2} step={0.01} {...form.getInputProps("total")} />
           </Group>
-          <Select
-            label="Legal entity (payer)"
-            required
-            data={(legalEntitiesQuery.data ?? []).map((le) => ({
-              value: le.id,
-              label: `${COUNTRY_FLAGS[le.country as Country] ?? ""} ${le.name}`,
-            }))}
-            value={form.values.legalEntityId}
-            onChange={(v) => form.setFieldValue("legalEntityId", v ?? "")}
-          />
           <Button type="submit" loading={mutation.isPending}>
             Save Changes
           </Button>
