+++ apps/web/src/routes/_authenticated/route.tsx
@@ -33,7 +33,6 @@ import {
   IconLogout,
   IconMoon,
   IconRefresh,
-  IconReportAnalytics,
   IconScissors,
   IconSettings,
   IconSun,
@@ -75,7 +74,6 @@ const LEGAL_ENTITY_TABS = [
   { value: "overview", label: "Overview", icon: IconLayoutDashboard },
   { value: "documents", label: "Documents", icon: IconFileText, badgeKey: "unassigned" as const },
   { value: "bank-accounts", label: "Bank accounts", icon: IconWallet },
-  { value: "reports", label: "Reports", icon: IconReportAnalytics },
   { value: "salons", label: "Salons", icon: IconScissors },
 ]
