+++ packages/db/src/schema/hair.ts
@@ -4,7 +4,6 @@ import { pgTable, integer, serial, text, timestamp, date } from "drizzle-orm/pg-
 import { appointment } from "./appointment"
 import { user } from "./auth"
 import { customer } from "./customer"
-import { legalEntity } from "./legal-entity"
 
 export const hairOrder = pgTable("hair_order", {
   id: text("id").primaryKey().$defaultFn(createId),
@@ -19,9 +18,6 @@ export const hairOrder = pgTable("hair_order", {
   customerId: text("customer_id")
     .notNull()
     .references(() => customer.id, { onDelete: "cascade" }),
-  legalEntityId: text("legal_entity_id")
-    .notNull()
-    .references(() => legalEntity.id, { onDelete: "restrict" }),
   createdById: text("created_by_id")
     .notNull()
     .references(() => user.id, { onDelete: "cascade" }),
